# chore(sqla): Address performance tradeoff with eager loading