@ktmud 2023-02-17
Thanks for the detailed explanation! Do you think it'd worth explicit set the lazy behavior even though it is the default?