@john-bodley 2023-02-17
@ktmud regarding,

> Do you think it'd worth explicit set the lazy behavior even though it is the default?

I'm kind of on the fence about this, and I did contemplate it. I think generally it's good/standard practice not to define functional arguments which are the default. Granted defaults can change, but it does/can result in code bloat.